requirements.txt
# Bus NER POC Dependencies

# Transformers & Deep Learning
transformers>=4.35.0
torch>=2.0.0
datasets>=2.14.0
accelerate>=0.24.0
seqeval>=1.2.2
python-dotenv>=1.0.0
sentencepiece>=0.2.0

# ONNX Runtime for optimized inference
optimum[onnxruntime]>=1.14.0
onnxruntime>=1.16.0

# API Framework
fastapi>=0.100.0
uvicorn>=0.23.0

# Data Validation
pydantic>=2.0.0

sentencepiece>=0.2.0

# python src/data_generator.py

# # Train model
# python src/train_ner.py --epochs 10

# # Export to ONNX
# python src/export_onnx.py --verify

# # Test inference
# python src/inference.py --query "AC sleeper from Bangalore to Mumbai"

# # Start API
# uvicorn src.api:app --host 0.0.0.0 --port 8000

# # Test API
# curl -X POST http://localhost:8000/ner/extract \
#   -H "Content-Type: application/json" \
#   -d '{"query": "VRL AC sleeper from Chennai to Hyderabad tomorrow"}'


# 1) Go to project
# cd /Users/int1964/NerModel_GTEXP1.0

# # 2) Create + activate virtual env
# python3 -m venv .venv
# source .venv/bin/activate

# # 3) Upgrade pip + install deps
# python -m pip install --upgrade pip
# pip install -r requirements.txt

# # 4) Generate training data
# python src/data_generator.py

# # 5) Train NEW model (save in a new folder name)
# python src/train_ner.py --epochs 10 --output models/bus_ner_transformer_v2

# # 6) Point API to this new model
# export BUS_NER_MODEL_PATH=/Users/int1964/NerModel_GTEXP1.0/models/bus_ner_transformer_v2

# # 7) Start server
# uvicorn src.api:app --host 0.0.0.0 --port 8000

# Run this:
# pip install numpypython src/train_ner.py --epochs 10 --output models/bus_ner_transformer_v2
# Better (recommended), install full project deps again so next missing package doesn’t fail one-by-one:
# pip install -r requirements.txtpip install numpypython src/train_ner.py --epochs 10 --output models/bus_ner_transformer_v2
# Why this happened:
# train_ner.py imports numpy
# current requirements.txt does not include numpy, so it wasn’t installed automatically.

#    python src/export_onnx.py --model models/bus_ner_transformer_v5 --output models/bus_ner_onnx_v5


# # 1. Generate failed patterns data
# cd /Users/int1964/NerModel_GTEXP1.0
# python src/generate_failed_data.py

# # This creates:
# # - data/failed_patterns_span.json (for debugging)
# # - data/failed_patterns_bio.json (for training)
# # Total: ~1,250 samples (25 templates × 50 each)

# # 2. Train on failed patterns
# python src/train_ner.py \
#     --model_path models/bus_ner_minilm_v7 \
#     --data_path data/failed_patterns_bio.json \
#     --output_dir models/bus_ner_minilm_v7_patched \
#     --num_train_epochs 15 \
#     --learning_rate 1e-5 \
#     --batch_size 4 \
#     --save_strategy epoch

# # 3. Test the patched model
# python src/inference.py \
#     --model_path models/bus_ner_minilm_v7_patched \
#     --query "show me buses from mumbai to delhi tomorrow morning"

